requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
environs[django]==9.5.0